+++ src/jamel/spheres/realSphere/IntegratedFactory.java
@@ -145,7 +145,7 @@ public class IntegratedFactory extends AbstractFactory {
 	 * @return 100.
 	 */
 	@Override
-	protected float getMaxLevelOfProduction() {
+	protected double getMaxLevelOfProduction() {
 		return 100;
 	}
